CSE loss with 1 layer lstm and dropout

+++ dataloaders/tc_dataloader.py
@@ -239,7 +239,7 @@ class TC_Dataloader(BaseDataset):
         # print(out)
         # print(label)
         # print(label.shape)
-        return out.float(), label.float()#.type(torch.LongTensor)
+        return out.float(), label.long()#.type(torch.LongTensor)
     
     def __len__(self):
         """
